+++ rollbar/contrib/flask/__init__.py
@@ -2,19 +2,18 @@
 Integration with Flask
 """
 
-import sys
-
 from flask import request
 import rollbar
 
 
 def report_exception(app, exception):
-    payload_data = {
-        'framework': 'Flask',
-    }
+    rollbar.report_exc_info(request=request)
 
-    if request.url_rule:
-        payload_data['context'] = str(request.url_rule)
 
-    rollbar.report_exc_info(sys.exc_info(), request, payload_data=payload_data)
-    
+def _hook(request, data):
+    data['framework'] = 'flask'
+
+    if request:
+        data['context'] = str(request.url_rule)
+
+rollbar.BASE_DATA_HOOK = _hook