+++ rollbar/contrib/pyramid/__init__.py
@@ -16,14 +16,7 @@ DEFAULT_WEB_BASE = 'https://rollbar.com'
 log = logging.getLogger(__name__)
 
 def handle_error(settings, request):
-    payload_data = None
-    try:
-        context = request.matched_route.name
-        payload_data = {'context': context}
-    except:
-        pass
-
-    rollbar.report_exc_info(sys.exc_info(), request, payload_data=payload_data)
+    rollbar.report_exc_info(sys.exc_info(), request)
 
 
 def parse_settings(settings):
@@ -118,6 +111,9 @@ def includeme(config):
         
         if request:
             request.environ['rollbar.uuid'] = data['uuid']
+
+            if request.matched_route:
+                data['context'] = request.matched_route.name
 
     rollbar.BASE_DATA_HOOK = hook